Add react-loadable and use it for the 2 chart components
