+++ app/components/DonutChart/__tests__/index-test.js
@@ -1,13 +1,9 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import DonutChart from '..';
+const mockedLoadable = jest.fn();
+jest.mock('utils/loadable', () => mockedLoadable);
 
-// mock nested components
-jest.mock('components/DonutChart/Path');
-jest.mock('components/Chart', () => 'div');
-jest.mock('components/Legend', () => 'div');
-
-it('renders correctly', () => {
-  const tree = renderer.create(<DonutChart dataLabel="test" dataKey="test" data={[]} />).toJSON();
-  expect(tree).toMatchSnapshot();
+it('uses loadable', done => {
+  import('../index').then(() => {
+    expect(mockedLoadable).toBeCalled();
+    done();
+  });
 });
