+++ flow-typed/npm/react-redux_v5.x.x.js
@@ -1,5 +1,5 @@
-// flow-typed signature: ba073ebd4a5e8e27ad3050bfe9f31c88
-// flow-typed version: d51977e735/react-redux_v5.x.x/flow_>=v0.63.0
+// flow-typed signature: 049afc81a36fe3ff0b8a977b916d4e33
+// flow-typed version: 35159486e5/react-redux_v5.x.x/flow_>=v0.63.0
 
 declare module "react-redux" {
   import type { ComponentType, ElementConfig } from 'react';
@@ -41,6 +41,7 @@ declare module "react-redux" {
   CP = Props for returned component
   Com = React Component
   ST = Static properties of Com
+  EFO = Extra factory options (used only in connectAdvanced)
   */
 
   declare type MapStateToProps<S: Object, SP: Object, RSP: Object> = (state: S, props: SP) => RSP;
@@ -64,6 +65,50 @@ declare module "react-redux" {
   |};
 
   declare type OmitDispatch<Component> = $Diff<Component, {dispatch?: Dispatch<*>}>;
+
+  declare type ConnectAdvancedOptions = {
+    getDisplayName?: (name: string) => string,
+    methodName?: string,
+    renderCountProp?: string,
+    shouldHandleStateChanges?: boolean,
+    storeKey?: string,
+    withRef?: boolean,
+  };
+
+  declare type SelectorFactoryOptions<Com> = {
+    getDisplayName: (name: string) => string,
+    methodName: string,
+    renderCountProp: ?string,
+    shouldHandleStateChanges: boolean,
+    storeKey: string,
+    withRef: boolean,
+    displayName: string,
+    wrappedComponentName: string,
+    WrappedComponent: Com,
+  };
+
+  declare type SelectorFactory<
+    Com: ComponentType<*>,
+    A,
+    S: Object,
+    OP: Object,
+    EFO: Object,
+    CP: Object
+  > = (dispatch: Dispatch<A>, factoryOptions: SelectorFactoryOptions<Com> & EFO) =>
+      MapStateToProps<S, OP, CP>;
+
+  declare export function connectAdvanced<
+    Com: ComponentType<*>,
+    A,
+    S: Object,
+    OP: Object,
+    CP: Object,
+    EFO: Object,
+    ST: {[_: $Keys<Com>]: any}
+    >(
+    selectorFactory: SelectorFactory<Com, A, S, OP, EFO, CP>,
+    connectAdvancedOptions: ?(ConnectAdvancedOptions & EFO),
+  ): (component: Com) => ComponentType<OP> & $Shape<ST>;
 
   declare export function connect<
     Com: ComponentType<*>,
@@ -211,5 +256,6 @@ declare module "react-redux" {
     Provider: typeof Provider,
     createProvider: typeof createProvider,
     connect: typeof connect,
+    connectAdvanced: typeof connectAdvanced,
   };
 }
