+++ flow-typed/npm/react-router-dom_v4.x.x.js
@@ -1,32 +1,32 @@
-// flow-typed signature: 928df3c83a85e46b1262c47094522f1f
-// flow-typed version: e9398ab068/react-router-dom_v4.x.x/flow_>=v0.63.x
+// flow-typed signature: 61446bc19a9bee479063b3cc256e235e
+// flow-typed version: 8840f4cb4b/react-router-dom_v4.x.x/flow_>=v0.63.x
 
 declare module "react-router-dom" {
-  import type { ComponentType, ElementConfig, Node } from 'react';
+  import type { ComponentType, ElementConfig, Node, Component } from 'react';
 
-  declare export class BrowserRouter extends React$Component<{|
+  declare export var BrowserRouter: Class<Component<{|
     basename?: string,
     forceRefresh?: boolean,
     getUserConfirmation?: GetUserConfirmation,
     keyLength?: number,
     children?: Node
-  |}> {}
+  |}>>
 
-  declare export class HashRouter extends React$Component<{|
+  declare export var HashRouter: Class<Component<{|
     basename?: string,
     getUserConfirmation?: GetUserConfirmation,
     hashType?: "slash" | "noslash" | "hashbang",
     children?: Node
-  |}> {}
+  |}>>
 
-  declare export class Link extends React$Component<{
+  declare export var Link: Class<Component<{
     className?: string,
     to: string | LocationShape,
     replace?: boolean,
     children?: Node
-  }> {}
+  }>>
 
-  declare export class NavLink extends React$Component<{
+  declare export var NavLink: Class<Component<{
     to: string | LocationShape,
     activeClassName?: string,
     className?: string,
@@ -36,7 +36,7 @@ declare module "react-router-dom" {
     children?: Node,
     exact?: boolean,
     strict?: boolean
-  }> {}
+  }>>
 
   // NOTE: Below are duplicated from react-router. If updating these, please
   // update the react-router and react-router-native types as well.
@@ -108,40 +108,40 @@ declare module "react-router-dom" {
     url?: string
   };
 
-  declare export class StaticRouter extends React$Component<{|
+  declare export var StaticRouter: Class<Component<{|
     basename?: string,
     location?: string | Location,
     context: StaticRouterContext,
     children?: Node
-  |}> {}
+  |}>>
 
-  declare export class MemoryRouter extends React$Component<{|
+  declare export var MemoryRouter: Class<Component<{|
     initialEntries?: Array<LocationShape | string>,
     initialIndex?: number,
     getUserConfirmation?: GetUserConfirmation,
     keyLength?: number,
     children?: Node
-  |}> {}
+  |}>>
 
-  declare export class Router extends React$Component<{|
+  declare export var Router: Class<Component<{|
     history: RouterHistory,
     children?: Node
-  |}> {}
+  |}>>
 
-  declare export class Prompt extends React$Component<{|
+  declare export var Prompt: Class<Component<{|
     message: string | ((location: Location) => string | boolean),
     when?: boolean
-  |}> {}
+  |}>>
 
-  declare export class Redirect extends React$Component<{|
+  declare export var Redirect: Class<Component<{|
     to: string | LocationShape,
     push?: boolean,
     from?: string,
     exact?: boolean,
     strict?: boolean
-  |}> {}
+  |}>>
 
-  declare export class Route extends React$Component<{|
+  declare export var Route: Class<Component<{|
     component?: ComponentType<*>,
     render?: (router: ContextRouter) => Node,
     children?: ComponentType<ContextRouter> | Node,
@@ -150,12 +150,12 @@ declare module "react-router-dom" {
     strict?: boolean,
     location?: LocationShape,
     sensitive?: boolean
-  |}> {}
+  |}>>
 
-  declare export class Switch extends React$Component<{|
+  declare export var Switch: Class<Component<{|
     children?: Node,
     location?: Location
-  |}> {}
+  |}>>
 
   declare export function withRouter<P: {}, Component: ComponentType<P>>(
     WrappedComponent: Component
