+++ package.json
@@ -90,6 +90,7 @@
     "react": "16.4.1",
     "react-dom": "16.4.1",
     "react-hot-loader": "^3.0.0-beta.7",
+    "react-loadable": "^5.5.0",
     "react-redux": "5.0.7",
     "react-router": "4.3.1",
     "react-router-dom": "4.3.1",
